requirements.txt
# necesario instalar tshark [sudo apt-get install tshark]
pyshark
python-logstash